+++ api/src/validators/query.validator.ts
@@ -1,6 +1,7 @@
 import { RequestHandler } from "express";
 import asyncHandler from "express-async-handler";
 import { checkSchema, Schema, validationResult } from "express-validator";
+import { BadRequestError } from "../errors/customErrors";
 
 const isInit = async (init: string) => {
   const initBoolean = init.toLocaleLowerCase() === "true";
@@ -30,7 +31,12 @@ const validateQuery = (schema: Schema): RequestHandler => {
       await checkSchema(schema, ["query"]).run(req);
       const errors = validationResult(req);
       if (!errors.isEmpty()) {
-        next({ status: "fail", code: 404, errors: errors.mapped() });
+        // next({ status: "fail", code: 404, errors: errors.mapped() });
+        throw new BadRequestError(
+          "Validation failed for request query",
+          404,
+          errors.mapped()
+        );
       }
 
       next();