+++ cms/src/hooks/useUserData.tsx
@@ -1,0 +1,5 @@
+const userUserData = () => {
+  const userData = () => {};
+
+  return userData;
+};